Extract BatchFileBuilder for readability.
