+++ library/src/main/java/com/novoda/downloadmanager/BatchFile.java
@@ -1,7 +1,5 @@
 package com.novoda.downloadmanager;
 
-import java.io.File;
-
 public class BatchFile {
 
     private final String networkAddress;
@@ -15,7 +13,7 @@ public class BatchFile {
     }
 
     static InternalBuilder from(DownloadBatchId downloadBatchId, String networkAddress) {
-        return new LiteFileBuilder(downloadBatchId, networkAddress);
+        return new BatchFileBuilder(downloadBatchId, networkAddress);
     }
 
     public String networkAddress() {
@@ -67,74 +65,4 @@ public class BatchFile {
                 + '}';
     }
 
-    public interface Builder {
-        Builder withIdentifier(DownloadFileId downloadFileId);
-
-        Builder saveTo(String path);
-
-        Builder saveTo(String path, String fileName);
-
-        Batch.Builder apply();
-    }
-
-    interface InternalBuilder extends Builder {
-        Builder withParentBuilder(Batch.InternalBuilder parentBuilder);
-    }
-
-    private static final class LiteFileBuilder implements InternalBuilder {
-
-        private final DownloadBatchId downloadBatchId;
-        private final String networkAddress;
-
-        private Optional<DownloadFileId> downloadFileId = Optional.absent();
-        private Optional<String> path = Optional.absent();
-
-        private Batch.InternalBuilder parentBuilder;
-
-        LiteFileBuilder(DownloadBatchId downloadBatchId, String networkAddress) {
-            this.downloadBatchId = downloadBatchId;
-            this.networkAddress = networkAddress;
-        }
-
-        @Override
-        public Builder withParentBuilder(Batch.InternalBuilder parentBuilder) {
-            this.parentBuilder = parentBuilder;
-            return this;
-        }
-
-        @Override
-        public Builder withIdentifier(DownloadFileId downloadFileId) {
-            this.downloadFileId = Optional.fromNullable(downloadFileId);
-            return this;
-        }
-
-        @Override
-        public Builder saveTo(String path) {
-            String networkAddressDerivedFileName = FileNameExtractor.extractFrom(networkAddress);
-            return saveTo(path, networkAddressDerivedFileName);
-        }
-
-        @Override
-        public Builder saveTo(String path, String fileName) {
-            if (path != null && fileName != null) {
-                this.path = Optional.of(path + fileName);
-            }
-
-            return this;
-        }
-
-        @Override
-        public Batch.Builder apply() {
-            String networkAddressDerivedFileName = FileNameExtractor.extractFrom(networkAddress);
-            String pathPrependedWithBatchId = prependBatchIdTo(path.or(networkAddressDerivedFileName), downloadBatchId);
-
-            parentBuilder.withFile(new BatchFile(networkAddress, downloadFileId, pathPrependedWithBatchId));
-            return parentBuilder;
-        }
-
-        private static String prependBatchIdTo(String filePath, DownloadBatchId downloadBatchId) {
-            return downloadBatchId.rawId() + File.separatorChar + filePath;
-        }
-
-    }
 }
