+++ library/src/main/java/com/novoda/downloadmanager/InternalBuilder.java
@@ -1,0 +1,5 @@
+package com.novoda.downloadmanager;
+
+interface InternalBuilder extends Builder {
+    Builder withParentBuilder(Batch.InternalBuilder parentBuilder);
+}
